refactored spook and possess
